Move user routes to their own file
